+++ Tests/MapboxMapsTests/MapView/Integration Tests/Map/MapViewIntegrationTests.swift
@@ -57,7 +57,7 @@ class MapViewIntegrationTests: IntegrationTestCase {
         XCTAssertEqual(originalFPS.rawValue, -1)
 
         let newFPS = 12
-        mapView.preferredFPS = .custom(fps: newFPS)
+        mapView.preferredFPS = PreferredFPS(rawValue: newFPS)
         XCTAssertNotEqual(originalFPS, mapView.preferredFPS)
         XCTAssertEqual(mapView.preferredFPS.rawValue, newFPS)
     }